ansible: update `upgrade-jar.yml` to use a variable
when the certs expired on the main CI we had to upgrade the jenkins
agents to use the newer certs. When I tried this on the macs it failed
due to a hard coded `/home` which doesn't exist on mac. Used the create
playbook as the inspiration for this change.